Prevent data loss when writes fail
When reading from Kafka, we now commit the read offset only after successfully writing to the Butler database.  This ensures that each message will be completely processed at least once.

This relies on the new `skip_existing` feature of `Butler.ingest` from DM-52179 to make Butler writes idempotent, so it is safe to retry the ingest even if it may have already happened in another process.